refactor: ativando extensão do portugol
